// src/services/rest-api/customer-service/model/emergency-contact-info-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Customer Service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.242
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface EmergencyContactInfoModel
 */
export interface EmergencyContactInfoModel {
    /**
     * 
     * @type {number}
     * @memberof EmergencyContactInfoModel
     */
    'emergencyContactId'?: number;
    /**
     * 
     * @type {string}
     * @memberof EmergencyContactInfoModel
     */
    'titleCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof EmergencyContactInfoModel
     */
    'titleName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof EmergencyContactInfoModel
     */
    'firstName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof EmergencyContactInfoModel
     */
    'lastName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof EmergencyContactInfoModel
     */
    'mobileNo'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof EmergencyContactInfoModel
     */
    'relationshipCode'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof EmergencyContactInfoModel
     */
    'relationshipDesc'?: string | null;
}